pyproject.toml
[project]
name = "etl-website-visits"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "paramiko>=4.0.0",
    "prefect>=3.4.24",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
]
